+++ R/compute-shape-grob.R
@@ -1,0 +1,109 @@
+
+#' Compute a shape grob on scale 0-1 npc for use for borders,
+#' target lines and mask
+#'
+#' @param from Stripped of surrounding `list()`
+#' @param shape "rect", "ellipse", "outline"
+#' @param data Data
+#' @param coord Coord object
+#' @param panel_params Opaque object
+#' @param expand Parameter for proportional expansion (not done yet)
+#'
+#' @return A [grid::grob()] object with units "npc" and data between 0 and
+#' 1 on screen scale. Coordinates have been transformed by `coord$transform(...)`.
+#' @noRd
+compute_shape_grob <- function (from, shape, data, coord, panel_params, expand) {
+  UseMethod("compute_shape_grob")
+  # if from is a grob, use it directly (after rescaling). If from is a data
+  # frame, make a grob
+  # and use that.
+  # Otherwise: if shape is rect or ellipse, then return those.
+  # If shape is outline then select points - a convex hull or sf grob
+}
+
+
+compute_shape_grob.grob <- function (from, shape, data, coord, panel_params,
+                                     expand) {
+  scale01 <- function (x) (x - min(x))/(max(x) - min(x))
+  from_cc <- allcoords(from)
+  # we don't transform for a raw grob; and we've expanded the bounding box
+  # already
+  x_scaled <- scale01(as.numeric(from_cc[, "x"]))
+  y_scaled <- scale01(as.numeric(from_cc[, "y"]))
+
+  grid::polygonGrob(x = grid::unit(x_scaled, "npc"),
+                    y = grid::unit(y_scaled, "npc"))
+}
+
+
+compute_shape_grob.data.frame <- function (from, shape, data, coord, panel_params,
+                                           expand) {
+  # this will be on scale of data
+  names(from) <- c("x", "y")
+  shape_cc <- coord$transform(from, panel_params)
+  from_grob <- polygonGrob(x = shape_cc$x, y = shape_cc$y,
+                            default.units = "native")
+  compute_shape_grob(from_grob, shape, data, coord, panel_params, expand)
+}
+
+
+compute_shape_grob.default <- function (from, shape, data, coord, panel_params,
+                                        expand) {
+  if (shape == "rect") {
+    grid::rectGrob()
+  } else if (shape == "ellipse") {
+    # resist the temptation to replace this with cirleGrob. You need to
+    # mess with it later.
+    gridExtra::ellipseGrob(x = 0.5, y = 0.5, size = 0.5, n = 180,
+                             position.units = "npc", size.units = "npc")
+  } else {
+    data <- subset_by_from(from, data)
+    # shape is "outline"
+    # we have to select points inside the data
+    # then build the hull or call geometry
+    if (! is.null(data$geometry) && inherits(data$geometry, "sfc")) {
+      rlang::check_installed("sf")
+      from <- sf::st_as_grob(data$geometry)
+      compute_shape_grob(from, shape, data, coord, panel_params, expand)
+    } else if ("x" %in% names(data) && "y" %in% names(data)){
+      # only here we use the actual x and y points
+      from_df <- hull_around(data$x, data$y, expand = 0)
+      from_df <- coord$transform(from_df, panel_params)
+
+      from_grob <- polygonGrob(x = from_df$x, y = from_df$y,
+                               default.units = "native")
+      compute_shape_grob(from_grob, shape, data, coord, panel_params, expand = 0)
+    } else {
+      cli::cli_warn(c("Couldn't find `x` and `y` to build convex hull.",
+                      "Falling back to shape = \"rect\""))
+      grid::rectGrob()
+    }
+  }
+}
+
+subset_by_from <- function (from, data) {
+  UseMethod("subset_by_from")
+}
+
+subset_by_from.logical <- function (from, data) {
+  data[from,]
+}
+
+subset_by_from.default <- function (from, data) {
+  if (is.null(names(from))) names(from) <- c("xmin", "xmax", "ymin", "ymax")
+  if (inherits(data$geometry, "sfc")) {
+    rlang::check_installed("sf")
+    bb <- sf::st_bbox(unlist(from))
+    bb <- sf::st_as_sfc(bb) # `crs =` doesn't work
+    bb <- sf::st_set_crs(bb, sf::st_crs(data$geometry))
+    in_bounds <- sf::st_within(data$geometry, bb)
+    in_bounds <- lengths(in_bounds) > 0
+    data[in_bounds,]
+  } else {
+    cond <- data$x >= from["xmin"] & data$x <= from["xmax"] &
+            data$y >= from["ymin"] & data$y <= from["ymax"]
+    data[cond,]
+  }
+}
+
+